LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_capture_mouse-expected.txt

Pointer Events capture test

Test Description: This test checks if setCapture/releaseCapture functions works properly. Complete the following actions:
Move your mouse over the black rectangle. pointermove event should be logged in the black rectangle
Move your mouse over the purple rectangle. pointerover event should be logged in the purple rectangle
Press and hold left mouse button over "Set Capture" button. "gotpointercapture" should be logged in the black rectangle
Move your mouse anywhere. pointermove should be logged in the black rectangle
Move your mouse over the purple rectangle. Nothig should happen
Move your mouse over the black rectangle. pointermove should be logged in the black rectangle
Release left mouse button. "lostpointercapture" should be logged in the black rectangle

Test passes if the proper behaviour of the events is observed.



Pointer Events Capture Test

The following pointer types were detected: mouse.

The following events were logged: pointermove@target0, pointermove@target1, gotpointercapture@target0, pointermove@target0, pointermove@target0, lostpointercapture@target0.


PASS gotpointercapture event received
PASS lostpointercapture event received
PASS pointerover event for black rectangle received
PASS pointerover event for purple rectangle received
PASS pointermove shouldn't trigger for the purple rectangle while the black rectangle has capture
PASS relatedTarget is null when the capture is set. relatedTarget is null
PASS pointermove received for captured element while out of it
PASS pointermove received for captured element while inside of it
